novel_analyzer/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.0.0
tqdm>=4.65.0
